Jiyoung/exception qna/review/찜/예약/판매 글 - 수정, 삭제 예외 처리 (#96)
* 리뷰+Qna 수정/삭제 예외처리

* 다른 사람이 등록한 공간 수정/삭제시 예외처리

* 예약 시스템 예외처리

* 찜 예외처리

* 필요없는 페이지 삭제

+++ TeamProject_SpaceRental/src/main/java/com/team/project/seller/controller/SellerController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.team.project.exception.NotUpdateException;
 import com.team.project.interceptor.Auth;
 import com.team.project.interceptor.Auth.Role;
 import com.team.project.qna.service.QnaService;
@@ -37,7 +38,8 @@ public class SellerController {
 		mView.setViewName("seller/spacelist");
 		return mView;
 	}
-
+	
+	@Auth(role = Role.SELLER)
 	@RequestMapping("/seller/sellerReview")
 	public ModelAndView sellerReview(ModelAndView mView, HttpServletRequest request) {
 		service.getReviewList(mView, request);
@@ -45,6 +47,7 @@ public class SellerController {
 		return mView;
 	}
 	
+	@Auth(role = Role.SELLER)
 	@RequestMapping("/seller/sellerQna")
 	public ModelAndView sellerQna(ModelAndView mView, HttpServletRequest request) {
 		service.getQnaList(mView, request);
@@ -72,8 +75,15 @@ public class SellerController {
 	@Auth(role = Role.SELLER)
 	@RequestMapping("/seller/spaceupdate")
 	public String spaceUpdate(HttpServletRequest request, HttpSession session, ModelAndView mView) {
-		service.getData(request);
 		if(session.getAttribute("id")!=null) {
+			SellerDto dto = service.getData(request);
+			int spaceOwner = dto.getUsers_num();
+			int user = service.getUsersNum(request, session);
+			
+			if(spaceOwner != user) {
+				throw new NotUpdateException("타인의 방을 수정하지 말아주세요.");
+			}
+			
 			usersService.getInfo(session, mView);
 		}
 		return "seller/spaceupdate";
@@ -83,16 +93,22 @@ public class SellerController {
 	@RequestMapping("/seller/update")
 	public String update(SellerDto dto, HttpServletRequest request, HttpSession session) {
 		service.update(dto, request);
-		System.out.println(dto.getmainImagePath());
-		System.out.println(dto.getSpace_name());
-		System.out.println(dto.getAddr());
 		
 		return "seller/update";		
 	}
 	
+	@Auth(role = Role.SELLER)
 	@RequestMapping("/seller/delete")
 	public String delete(int space_num, HttpServletRequest request) {
-		service.delete(space_num, request);
+		SellerDto dto = service.getData(request);
+		int spaceOwner = dto.getUsers_num();
+		int user = service.getUsersNum(request, request.getSession());
+		
+		if(spaceOwner != user) {
+			throw new NotUpdateException("타인의 방을 삭제하지 말아주세요.");
+		}
+		request.setAttribute("space_num", space_num);
+		service.delete(request);
 		qnaService.deleteContent2(space_num, request);
 		return "redirect:/seller/spacelist"; //다시 한번 물어보는 것으로 수정예정
 	}
